legacy/requirements.txt
matplotlib==3.7.1
numpy==1.25.0
pandas==1.5.3
scikit-learn==1.2.2
scipy==1.10.1
torch==1.13.1
torch_ema==0.3
tensorboard
tqdm
pyyaml
